Implement warnTheSheep function
Implemented the warnTheSheep function to warn the sheep in front of the wolf about the impending danger. The function calculates the positions of the wolf and the sheep in the queue and provides the appropriate warning message. If the wolf is at the front, it returns "Pls go away and stop eating my sheep," otherwise, it returns "Oi! Sheep number N! You are about to be eaten by a wolf!" where N is the sheep's position.